requirements.txt
Pillow==9.4.0
pystray==0.19.4
